Set up javadocs build with exclusions.
Command implementations will be added to a new package,
new nu.nerd.entitymeta.commands, which I want to exclude
from JavaDoc processing. Also, to implement configuration
reloading, Configuration class now needs to be public
since it is accessed from the commands package. So
Configuration.java needs to be excluded too.

The JavaDoc FAQ lists several unsatisfactory options:
http://www.oracle.com/technetwork/java/javase/documentation/index-137483.html#exclude

Java 1.9 may be adding an @exclude tag which could then be applied to the
Configuration class and EntityMeta.CONFIG. Which would be nice.